% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{Print methods for \pkg{schoenberg}}
\arguments{
\item{x}{Object to be printed (object is used to select a method).}

\item{...}{Additional arguments.}
}
\description{
Print methods for \pkg{schoenberg} output objects with classes exported from \pkg{schoenberg}.
}
