% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoenberg.R
\name{print.schoenberg}
\alias{print.schoenberg}
\title{print method for objects of the "schoenberg" class.}
\usage{
\method{print}{schoenberg}(x, ...)
}
\arguments{
\item{x}{Object to be printed.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Printed results from objects of the "schoenberg" class.
}
\description{
print method for objects of the "schoenberg" class.
}
