% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoenberg.R
\name{rekey}
\alias{rekey}
\title{Re-express a "schoenberg" class object with a different leading tone or accidental.}
\usage{
rekey(tone_mat, lead_tone = NULL, use_sharps = TRUE)
}
\arguments{
\item{tone_mat}{Object of the class "schoenberg" produced by the \code{schoenberg()} function.}

\item{lead_tone}{Name of the note to use as the leading tone of the matrix.}

\item{use_sharps}{Logical scalar that determines whether accidentals should be represented as sharps (\code{TRUE}; default) or flats (\code{FALSE}).}
}
\value{
A 12-tone matrix of the "schoenberg" class.
}
\description{
Re-express a "schoenberg" class object with a different leading tone or accidental.
}
\examples{
## Let's create a vector of notes to use in creating our inital 'tone_mat' matrix:
prime01 <- c("F#", "F", "D", "E", "D#", "C", "A", "C#", "G#", "B", "A#", "G")
tone_mat <- schoenberg(prime0 = prime01)

# Now, let's change the leading tone to "C":
rekey(tone_mat = tone_mat, lead_tone = "C")

# And let's also change the accidentals to flats:
rekey(tone_mat = tone_mat, lead_tone = "C", use_sharps = FALSE)
}
